rework README completely, makefile and configuration aspects
* rewrite README, try to help DevOps deploy this project
* delete env from Makefile
* combine mini-micro-tiny configuration/*.go files into one
